# go/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  internal_handlers.HealthResponse:
    properties:
      server:
        $ref: '#/definitions/internal_handlers.ServiceStatus'
      services:
        items:
          $ref: '#/definitions/internal_handlers.ServiceStatus'
        type: array
      status:
        description: '"UP" or "DEGRADED" or "DOWN"'
        type: string
    type: object
  internal_handlers.ServiceStatus:
    properties:
      latency:
        description: Will be omitted if empty
        type: string
      message:
        description: Will be omitted if empty
        type: string
      name:
        type: string
      status:
        type: string
      url:
        description: Will be omitted if empty
        type: string
    type: object
  neneloga_internal_models.Log:
    properties:
      created_at:
        type: string
      id:
        type: integer
      level:
        type: string
      message:
        type: string
      serverID:
        type: integer
      updated_at:
        type: string
      userID:
        description: nullable
        type: integer
    type: object
  neneloga_internal_models.Server:
    properties:
      created_at:
        type: string
      hostname:
        type: string
      id:
        type: integer
      ipaddress:
        type: string
      logs:
        items:
          $ref: '#/definitions/neneloga_internal_models.Log'
        type: array
      name:
        type: string
      os:
        type: string
      status:
        $ref: '#/definitions/neneloga_internal_models.ServerStatus'
      updated_at:
        type: string
      userID:
        type: integer
    type: object
  neneloga_internal_models.ServerStatus:
    enum:
    - active
    - deleted
    - offline
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusDeleted
    - StatusOffline
  neneloga_internal_models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        description: admin, user
        type: string
      servers:
        items:
          $ref: '#/definitions/neneloga_internal_models.Server'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for Neneloga.
  title: Neneloga API
  version: "1.0"
paths:
  /:
    get:
      description: Get a list of all available endpoints
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Home endpoint
      tags:
      - system
  /chuck:
    get:
      description: Get a random Chuck Norris joke
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Chuck Norris Joke
      tags:
      - fun
  /health:
    get:
      description: Health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.HealthResponse'
      summary: Health
      tags:
      - health
  /servers:
    get:
      description: Get a list of servers with filtering, sorting, and pagination
      parameters:
      - description: 'Filter by status: active, deleted, offline'
        in: query
        name: status
        type: string
      - description: Filter by OS
        in: query
        name: os
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Search in name, hostname, IP
        in: query
        name: search
        type: string
      - description: Sort by field (prefix with - for DESC)
        enum:
        - -created_at
        - created_at
        - -name
        - name
        - -status
        - status
        in: query
        name: sort
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all servers
      tags:
      - servers
    post:
      consumes:
      - application/json
      description: Add a new server to the database
      parameters:
      - description: Server object
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/neneloga_internal_models.Server'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/neneloga_internal_models.Server'
      summary: Create a new server
      tags:
      - servers
  /servers/{id}:
    delete:
      description: Change server status to 'deleted' instead of hard delete
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      summary: Soft delete a server
      tags:
      - servers
    get:
      description: Get server by ID with its logs
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/neneloga_internal_models.Server'
      summary: Get a single server
      tags:
      - servers
    put:
      consumes:
      - application/json
      description: Update an existing server
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: Server object
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/neneloga_internal_models.Server'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/neneloga_internal_models.Server'
      summary: Update a server
      tags:
      - servers
  /users/:
    get:
      description: Get a list of all users and their logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/neneloga_internal_models.User'
            type: array
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/neneloga_internal_models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/neneloga_internal_models.User'
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      description: Get a single user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/neneloga_internal_models.User'
      summary: Get a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/neneloga_internal_models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/neneloga_internal_models.User'
      summary: Update a user
      tags:
      - users
swagger: "2.0"